+++ src/libs/Detach/CHANGELOG.md
@@ -10,11 +10,16 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added `Geometry3D.SpherePyramid` and `Geometry3D.SphereCastPyramid` methods.
 - Added `GetCorners` method to `ViewFrustum` (previously `Frustum`) struct.
 - Added support for `bool`, `Vector2`, `Vector3`, `Vector4`, and `Quaternion` to `InlineInterpolatedStringHandlerUtf8`.
+- Added `MathUtils.IsAlmostZero` method.
 
 ### Changed
 
 - Changed the order of parameters in some `Geometry3D.ClosestPointIn...` methods. The `point` parameter is now always the first parameter.
 - Renamed `Frustum` to `ViewFrustum`.
+
+### Fixed
+
+- Fixed minor precision issues in `Geometry2D.LineLine` and `Geometry2D.LineRectangle` methods.
 
 ## 0.24.3